+++ packages/core/src/components/sections/ScrollToTopButton/ScrollToTopButton.tsx
@@ -2,7 +2,7 @@ import Icon from 'src/components/ui/Icon'
 
 import Button from '../../ui/Button'
 import Section from '../Section'
-import type { UIButtonProps } from '../../ui/Button'
+import type { ButtonProps } from '../../ui/Button'
 
 interface ScrollToTopButtonProps {
   /**
@@ -14,12 +14,12 @@ interface ScrollToTopButtonProps {
    * Button's icon.
    * @default <Icon name="CaretUp" width={16} height={16} weight="bold" />
    */
-  icon?: UIButtonProps['icon']
+  icon?: ButtonProps['icon']
   /**
    * Button icon's position.
    * @default 'left'
    */
-  iconPosition?: UIButtonProps['iconPosition']
+  iconPosition?: ButtonProps['iconPosition']
 }
 
 function ScrollToTopButton({