+++ packages/core/src/components/ui/Button/Button.tsx
@@ -1,20 +1,44 @@
 import { Button as UIButton, Icon as UIIcon } from '@faststore/ui'
-import type { ReactNode } from 'react'
-import type { ButtonProps } from '@faststore/ui'
+import type { ReactNode, AriaAttributes } from 'react'
+import type { ButtonProps as UIButtonProps } from '@faststore/ui'
+
+import styles from './button.module.scss'
 
 export type Variant = 'primary' | 'secondary' | 'tertiary'
 export type Size = 'small' | 'regular'
 export type IconPosition = 'left' | 'right'
 
-export type UIButtonProps = {
+export interface ButtonProps
+  extends Omit<UIButtonProps, 'aria-label' | 'disabled'> {
+  /**
+   * Specifies the component variant
+   */
   variant?: Variant
+  /**
+   * Specifies the size variant
+   */
   size?: Size
+  /**
+   * Defines use of inverted colors
+   */
   inverse?: boolean
+  /**
+   * Icon component
+   */
   icon?: ReactNode
+  /**
+   * Specifies where the icon should be positioned
+   */
   iconPosition?: IconPosition
+  /**
+   * For accessibility purposes, defines an ARIA label to the element when no label is provided
+   */
+  'aria-label'?: AriaAttributes['aria-label']
+  /**
+   * Specifies that this button should be disabled
+   */
+  disabled?: boolean
 }
-
-type Props = ButtonProps & UIButtonProps
 
 function Button({
   variant,
@@ -23,14 +47,19 @@ function Button({
   icon,
   iconPosition,
   children,
+  'aria-label': ariaLabel,
+  disabled,
   ...props
-}: Props) {
+}: ButtonProps) {
   return (
     <UIButton
+      aria-label={ariaLabel}
+      className={styles.fsButton}
       data-fs-button
-      data-fs-button-variant={variant}
-      data-fs-button-size={size}
       data-fs-button-inverse={inverse}
+      data-fs-button-size={size}
+      data-fs-button-variant={variant}
+      disabled={disabled}
       {...props}
     >
       {iconPosition === 'left' && <UIIcon component={icon} />}