+++ packages/core/src/components/ui/Button/ButtonLink/ButtonSignIn/ButtonSignIn.tsx
@@ -1,7 +1,8 @@
 import { useSession } from '@faststore/sdk'
 
-import { ButtonLink } from 'src/components/ui/Button'
 import Icon from 'src/components/ui/Icon'
+import { ButtonLink } from 'src/components/ui/Button'
+import styles from 'src/components/ui/Button/button.module.scss'
 
 const ButtonSignIn = () => {
   const { person } = useSession()
@@ -10,7 +11,7 @@ const ButtonSignIn = () => {
     <ButtonLink
       data-fs-button-signin-link
       href={person?.id ? '/account' : '/login'}
-      className="text__title-mini signin-link"
+      className={`${styles.fsButton} text__title-mini signin-link`}
       variant="tertiary"
     >
       <Icon name="User" width={18} height={18} weight="bold" />